Separate implementation of MouseMode from CheckableTool...
MouseModes and CheckableTools are really completely independent
concepts, and so their implementations have been untied so that they can
be used separately.
